fully implemented

+++ src/Databases/Interfaces/BaseDatabase.h
@@ -1,14 +1,10 @@
-//
-// Created by jorge on 15/11/2022.
-//
-
 #ifndef KING_TEST_BASEDATABASE_H
 #define KING_TEST_BASEDATABASE_H
 
 
 class BaseDatabase {
 public:
-    virtual int getTotalEntries() = 0;
+    virtual unsigned long long getTotalEntries() = 0;
 };
